notes refactoring folders structure
