+++ src/widgets/news/News.tsx
