+++ src/widgets/notes/Notes.tsx
