+++ src/widgets/notes/components/notesForCity/NotesForCity.tsx
