+++ src/widgets/notes/components/notesWithoutCity/NotesWithoutCity.tsx
