+++ src/widgets/weather/Weather.tsx
@@ -1,7 +1,7 @@
 import React, {useEffect, useState} from 'react';
 import {getWeather, WeatherAPIType} from '../../api/weatherAPI';
 import {ErrorMessage} from '../../common/errorMessage/ErrorMessage';
-import {InformationOfCity} from '../informationOfCity/InformationOfCity';
+import {InformationOfCity} from './components/informationOfCity/InformationOfCity';
 
 type PropsType = {
     city?: string;
