+++ src/widgets/weather/components/informationOfCity/InformationOfCity.module.css
